added factorial problem

+++ c++ basic_for_loop_problems/factorial.c++
@@ -1,9 +1,9 @@
 #include <iostream>
 using namespace std;
 
-int fact(int n)
+long long int fact(long long int n)
 {
-     int fact = 1;
+     long long int fact = 1;
      for (int i = 1; i <= n; i++)
      {
           fact = fact * i;
@@ -15,6 +15,6 @@ int main()
 {
      int n;
      cin >> n;
-     int res = fact(n);
+     long long int res = fact(n);
      cout << "Fact of " << n << ": " << res;
 }
